Move Subscription identifiers to Subscription class
